requirements.txt
alembic==1.13.1
aniso8601==9.0.1
asttokens==2.4.1
blinker==1.8.1
click==8.1.7
decorator==5.1.1
executing==2.0.1
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
ipython==8.24.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
Mako==1.3.3
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.18.0
PyJWT==2.8.0
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.29
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.11.0
wcwidth==0.2.13
Werkzeug==3.0.2
